Refactor CollectionModal into separate components
